Start to strip out minimal, Linux only, FFMpeg only
