--- .travis.yml
@@ -1,8 +1,0 @@
-language: ruby
-rvm:
-  - 1.9.3
-  - 2.0.0
-  - 2.1.2
-before_install:
-  - sudo apt-get update -qq
-  - sudo apt-get install -y libav-tools
