+++ README.md
@@ -1,14 +1,8 @@
 # AV
 
-A Ruby Programmable interface for FFmpeg and Libav.
+A experimental Ruby Programmable interface for FFmpeg. This is *not* a general
+replacement for the upstream parent Gem!
 
-## Status
-
-[![Build Status](https://travis-ci.org/ruby-av/av.svg?branch=master)](https://travis-ci.org/ruby-av/av)
-[![Coverage Status](https://coveralls.io/repos/ruby-av/av/badge.png?branch=master)](https://coveralls.io/r/ruby-av/av?branch=master)
-[![Code Climate](https://codeclimate.com/github/ruby-av/av/badges/gpa.svg)](https://codeclimate.com/github/ruby-av/av)
-[![Dependency Status](https://gemnasium.com/ruby-av/av.svg)](https://gemnasium.com/ruby-av/av)
-[![Gem Version](https://badge.fury.io/rb/av.svg)](http://badge.fury.io/rb/av)
 ## Installation
 
 Add this line to your application's Gemfile:
