+++ Rakefile
@@ -1,6 +1,8 @@
-require "bundler/gem_tasks"
-require "rake"
-require "rspec/core/rake_task"
+# frozen_string_literal: true
+
+require 'bundler/gem_tasks'
+require 'rake'
+require 'rspec/core/rake_task'
 
 RSpec::Core::RakeTask.new(:spec)
 task test: :spec
