+++ lib/av/engine.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Av
   class Engine < ::Rails::Engine
     initializer 'av.logger' do
