+++ lib/av/logger.rb
@@ -1,17 +1,19 @@
-require "logger"
+# frozen_string_literal: true
+
+require 'logger'
 
 module Av
   module Logger
-    def log message
-      logger.info "[AV] #{message}" if options[:log]
+    def log(message)
+      logger.info("[AV] #{message}") if options[:log]
     end
-  
+
     def logger
-      @logger ||= ::Logger.new(STDOUT)
+      @logger ||= ::Logger.new($stdout)
     end
 
     def logger=(logger)
       @logger = logger
     end
-  end  
+  end
 end
