+++ lib/av/param_hash.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 module Av
   class ParamHash < Hash
     def to_s
       line = []
-      self.each do |option, value|
+      each do |option, value|
         value = value.join(',') if value.is_a?(Array)
         line << "-#{option} #{value}"
       end
