+++ lib/av/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Av
-  VERSION = "0.9.0"
+  VERSION = '1.0.0'
 end
